feat: 컴포넌트 분리 및 영양리포트섹션 퍼블리싱
